# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Aurélien",
        "description": "C très Bo Whoulla",
        "contact": {
            "email": "admin9@fromage.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/authors": {
            "get": {
                "tags": [
                    "CRUD_Author"
                ],
                "summary": "Get all authors",
                "description": "Returns all authors",
                "operationId": "getAllAuthor",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "post": {
                "tags": [
                    "CRUD_Author"
                ],
                "summary": "Create author",
                "description": "create author",
                "operationId": "createAuthor",
                "requestBody": {
                    "description": "Pass author information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Store a newly created resource in storage.",
                                        "type": "string",
                                        "example": "Rage vince"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/authors/{id}": {
            "get": {
                "tags": [
                    "CRUD_Author"
                ],
                "summary": "Get author by ID",
                "description": "Returns author by his id",
                "operationId": "getAuthorByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "author id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Id not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CRUD_Author"
                ],
                "summary": "Delete author by ID",
                "description": "Delete author by his id",
                "operationId": "deleteAuthorByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "book id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "authors was deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "patch": {
                "tags": [
                    "CRUD_Author"
                ],
                "summary": "Update author by ID",
                "description": "Update author by his id",
                "operationId": "updateAuthorByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "author id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass author information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Update the specified resource in storage.",
                                        "type": "string",
                                        "example": "Laisse Béton update"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Id not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        []
    ]
}